Connection abort issues before the client-to-squid TLS connection established
This patch try to account the current bumping state before the
client-to-squid established. Currently defines the following states:
    bumpStateExpectTlsHandshake: Before the client hello message received.
    bumpStateParsingTlsHandshake: While parses the client hello message
    bumpStateParsingDone: The client hello message parsed
    bumpStatePeekEvaluate: After peeks at client, run step2 acls evaluate
                           client handshake with openSSL
    bumpStatePeekAtServer: Wait the server side peeking procedure to
                           be finished.
    bumpStateGenerateContext: While generates internal squid structures
                              and generate certificates to establish TLS
                              connection with client.
    bumpStateTlsNegotiate: TLS Negotiation with client
    bumpStateTlsEstablish: The TLS connection established

If TLS bumping abort for a reason before the client-to-squid connection
established, squid generates a fake ClientHttpRequest using as uri
and error in the form "Error-bump-aborted:[Bumping-State]" and add it to
ConnStateData::pipeline to be logged.

The patch also removes the ConnStateData::parsingTlsHandshake member and
replace it with a method.

TODO:  remove the ConnStateData::switchedToHttps_  and use the replaced method
       ConnStateData::switchedToHttps() {return  bumpingState == Ssl::bumpStateTlsEstablish; }
       instead.

TODO: Maybe more detailed logging description of the related errors.