Create new ClientHttpRequest object when splices at step3
The reason is to force produce a second logged line in access log when a
peeked conenction spliced at step3 as we do when the connection spliced at
step2 or after splicing because of unsupported protocol.

Still there is an inconsistency because the generated client request after
splicing at step2 passes callouts but after splicing at step3 this is not
happens.

To fix this inconsistency we should drop switchToTunnel and use a variant
of ConnStateData::splice method