Investigate the ConnStateData::detailErrorFailures
This template method designed while tried to provide more details
using c++ exceptions

Finally this is used to catch exceptions (and provide them as details)
while trying to splice after bump or tunnel to server after an error
(on_unsupported_protocol feature). The failures in these cases can only
be caused because of bugs or unexpected situations. So the helpfulness
of the new template method is not very clear for now.

The goal originaly was to replace the ERR_DETAIL_SSL_BUMP_SPLICE error
detail with a set with new details. This goal is not achieved
and this err detail is now used as a general error detail if the
splice is failed without providing any detail.

The patch also adds the ERR_DETAIL_TUNNEL_ON_ERROR as a general error
if tunnling after an error failed because an unknonw reason.

I am expecting that the ERR_DETAIL_SSL_BUMP_SPLICE and the new
ERR_DETAIL_TUNNEL_ON_ERROR are in practice unused and maybe should
replaced with a common ERR_DETAIL_UNKNOWN_REASON.

This patch also modifies the splice, initiateTunneledRequest and
fakeAConnectRequest ConnStateData methods to be void methods
instead of returning a boolean value. In practice always returns
true except something very strange (bug?) happens.