Fixed FTP error detailing. Changed FTP error extension headers/lines.
The official code incorrectly assumed that the error detail is always an
errno. That was a known problem.

    500-Error: (1) Operation not permitted

The earlier branch code reported (often boring/uninformative and always
human-unfriendly) detail "type" (e.g., 23) together with non-portable
and human-unfriendly logCode() (e.g., SYSERR=1).

If we take a system call error as an example, Squid should be sending
both the errno (for automation, especially where portability is not a
concern) and the corresponding error message (for the convenience of
human-driven FTP clients and sysadmins). Squid now does that, using two
FTP response header lines for easier parsing.

    500-Error-Detail-Brief: SYSERR=1
    500-Error-Detail-Verbose: Operation not permitted