Do not use exceptions to report tlsAttemptHandshake() failures
Exceptions are great for reporting function failures _through_ callers
that do not want to deal with the problem (immediately). Existing
tlsAttemptHandshake() callers deal with the problem immediately.

Throwing highly customized exceptions outside std::exception hierarchy
is dangerous because only the specialized catcher can handle them, but
the compiler cannot detect unhandled exceptions.

Throwing smart pointers to hierarchical objects is error-prone because
the catcher must guess the exact type of the underlying object in order
to catch that exception. Catching a Parent reference will catch a Child
reference but catching a RefCount<Parent> reference does not catch a
RefCount<Child> reference!

Finally, addressing the existing TODO about (not) calling SetSelect when
parsing already received Hello is likely to require further adjustments
that may make tlsAttemptHandshake() itself unnecessary. The underlying
Security::Accept() does not throw, in part, because it cannot tell for
sure which of the result outcomes should be treated as errors from the
caller point of view.