detailError() before calling clientTunnelOnError()
clientTunnelOnError() does not know enough about the error to detail it.
That detailing should be the responsibility of the code that discovered
the error, not the code that attempts to minimize its impact.

Most other commit changes deal with propagating err_type information
along with the ErrorDetail to access log. TODO: Find a way to make it
easier to treat both pieces of information as one entity.

Also added questions and concerns for future work.

XXX: This code does not compile because some of the error fields are
inaccessible in new contexts. Addressing the above TODO should help.