Make Must() ID search case-insensitive and do not require 0x prefixes
Squid logs exception IDs using small HEX letters and without 0x: (e.g.,
ERR_ACCESS_DENIED/EXCEPTION=18686e32).

We could change Squid to log using capital letters. It is a matter of
taste. HEX numbers with small letters feel a bit more readable to me.

We could change Squid to log with the 0x prefix, but that feels wasteful
because the user is not supposed to interpret these IDs as HEX numbers
and the (meaningless-to-the-user) 0x prefix will be repeated for all
such log entries. Moreover, we may introduce other ID (spaces) that are
not HEX. We can do that using a custom non-HEX prefix letter, but if log
processing software starts to expect 0x prefixes, such an expansion
would be more painful and/or look less natural.