Removed all ERR_DETAIL_* globals
Since ErrorCode values are often reused in different circumstances, and
triage often requires pinpointing the circumstances that led to a
particular error, each EnumeratedErrorDetail object should be used in
exactly one location, tying that code to the error being triaged.

It is wrong to declare a _global_ that should be used in a single
location, especially when that global is a fairly complex object and not
just an enum value:

* resources are wasted on errors that do not get triggered
* error detail documentation is separated from the error context
* naturally increases developer resistance to detailing errors because
  details requires fiddling with the global declarations in some files
  that the developers are not "working on" and are not familiar with.

The only (small) advantage of global declarations is that the compiler
ensures object name uniqueness. Still, due to typos, error detail names
could be mistyped or even be duplicated. For example, while working on
this commit, I discovered that we left "ERR_DETAIL_" prefix in three
ERR_DETAIL_UNEXPECTED_* names (now fixed).

TODO: Detect accidentally duplicated error detail names (a lint check).

Also added a missing "break" to the Security::IoResult::ioSuccess case
in ConnStateData::handleSslBumpHandshakeError(). The bug probably had no
observable effect because the second (incorrect) error detail does not
overwrite the first (correct) one.