Fix Security::ErrorDetail::brief to log human-friendly name for GnuTLS
When GnuTLS is used the ErrorDetail::ioErrorNo is in practice the only
wahy to retrieve a usefull for users information about the reason of
a failed connection. Moreover the GnuTls provides an easy to use
function which converts from an integer error code to error name.
This patch fixes Security::ErrorDetail::brief to print error name
when GnuTls is used.

For OpenSSL the most important information is included in
ErrorDetail::lib_error_no code which is currently logged as an integer
code